fucking histidine shit
YOU SAY IT DELIBERATELY TO PISS ME OFF.
I'm respectful enough to not talk to you about smell, alcohol, Futurama, Jurassic Park, reincarnation, or time travel,
and all I ask in return is to respect that there's one word that sounds angry whenever I hear it.
Why does this shit have to be so one sided!